autoadd

+++ hw/hw4/mytar/header.h
@@ -58,7 +58,7 @@ typedef struct header{
 /*=============UTILITY FUNCTIONS=================================*/
 void print_err(char *msg);
 uint64_t hash_fieldHeader(uint8_t *field, int LENGTH);
-void dec_to_oct_asciiString(uint64_t *buff, unsigned long value ,int LENGTH);
+void dec_to_oct_asciiString(uint8_t *buff, unsigned long value ,int LENGTH);
 int insert_special_int(char *where, size_t size, int32_t val);
 /*=================================================================*/
 /*=================HEADER FUNCTIONS================================*/
